move some non-framework items out of `chia.cmds.cmd_classes` (#19047)
* move some non-framework items out of `chia.cmds.cmd_classes`

* move the other new stuff too
